requirements.txt
discord.py>=2.0
PyYAML
python-dotenv
